Moved PassThroughResponseErrorHandler to public interface
