+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>org.zalando</groupId>
     <artifactId>riptide</artifactId>
-    <version>0.1.0-SNAPSHOT</version>
+    <version>0.2.0-SNAPSHOT</version>
 
     <organization>
         <name>Zalando SE</name>
