+++ src/test/java/org/zalando/riptide/AnyDispatchTest.java
@@ -25,6 +25,7 @@ import org.springframework.core.io.ClassPathResource;
 import org.springframework.http.client.ClientHttpResponse;
 import org.springframework.test.web.client.MockRestServiceServer;
 import org.springframework.web.client.RestTemplate;
+import org.zalando.riptide.handler.PassThroughResponseErrorHandler;
 import org.zalando.riptide.model.AccountBody;
 
 import java.io.IOException;
