+++ src/test/java/org/zalando/riptide/CaptureTest.java
@@ -29,6 +29,7 @@ import org.springframework.http.client.ClientHttpResponse;
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
 import org.springframework.test.web.client.MockRestServiceServer;
 import org.springframework.web.client.RestTemplate;
+import org.zalando.riptide.handler.PassThroughResponseErrorHandler;
 import org.zalando.riptide.model.Account;
 import org.zalando.riptide.model.AccountBody;
