+++ src/test/java/org/zalando/riptide/ContentTypeDispatchTest.java
@@ -27,6 +27,7 @@ import org.springframework.http.client.ClientHttpResponse;
 import org.springframework.http.converter.json.MappingJackson2HttpMessageConverter;
 import org.springframework.test.web.client.MockRestServiceServer;
 import org.springframework.web.client.RestTemplate;
+import org.zalando.riptide.handler.PassThroughResponseErrorHandler;
 import org.zalando.riptide.model.Error;
 import org.zalando.riptide.model.Problem;
 import org.zalando.riptide.model.Success;
